pyproject.toml
[tool.poetry]
name = "flax"
version = "0.1.0"
description = "simulate trade system!"
authors = ["agari"]

[tool.poetry.dependencies]
python = "^3.7"
pydantic = {extras = ["dotenv"], version = "^1.4"}
sqlalchemy = "^1.3.15"
psycopg2 = "^2.8.4"
sqlalchemy-stubs = "^0.3"
fastapi = "^0.52.0"
uvicorn = "^0.11.3"
pyjwt = "^1.7.1"
passlib = {extras = ["bcrypt"], version = "^1.7.2"}
python-multipart = "^0.0.5"
databases = {extras = ["postgresql"], version = "^0.2.6"}

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.770"
black = "^19.10b0"
flake8 = "^3.7.9"
ipykernel = "^5.1.4"
alembic = "^1.4.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
